@@ -5,16 +5,16 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.1.4.RELEASE</version>
+        <version>2.7.5</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>uz.onlinestor</groupId>
     <artifactId>onlinestoreback</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <name>onlinestoreback</name>
-    <description>Online store</description>
+    <description>Online store back</description>
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>11</java.version>
         <jsonwebtoken.version>0.8.0</jsonwebtoken.version>
     </properties>
     <dependencies>
@@ -70,8 +70,12 @@
         <dependency>
             <groupId>org.hibernate</groupId>
             <artifactId>hibernate-core</artifactId>
-            <version>6.1.6.Final</version>
+            <version>6.2.0.Final</version>
             <type>pom</type>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-data-jdbc</artifactId>
         </dependency>
 
     </dependencies>
